use correct tx price color and fix tx refresh button layout
