+++ src/renderer/scss/component/_table.scss
@@ -109,8 +109,4 @@
   td:nth-of-type(5) {
     width: 15%;
   }
-
-  .badge {
-    background-color: transparent !important;
-  }
 }
