+++ src/renderer/types/transaction.js
@@ -1,0 +1,11 @@
+// @flow
+export type Transaction = {
+  amount: number,
+  claim_id: string,
+  claim_name: string,
+  fee: number,
+  nout: number,
+  txid: string,
+  type: string,
+  date: Date,
+};
